Add routing functions
The path route type maps a regular expression to one or more functions
that can find a resource. Subexpressions in the regex are passed in as
arguments, and the find functions are tried in order, stopping when one
returns a non-null resource.

The resource type remains abstract at this point.