Add logic to handle request and response types
When methods are defined with content type readers and/or writers then
the correct ones are selected based on the request's entity type and
accept type (i.e., the media types specified in the Content-Type and
Accept headers).

When no matching reader or writer is found then the response fails with
the appropriate status (406 or 415) and the list of permitted types in
the body.