Add test endpoint that takes a path argument
GET /evens/{n} returns HTML with the value of n in the body for even n
and 404's otherwise. This revealed some fuzziness in what exactly was
being found by the "find resource" behavior, the upshot being the
introduction of a found_resource structure that links to the resource
definition as well as an optional blob of arbitrary data that can be
used in the respond method.